% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my_allele.freq.R
\name{my_allele_freq}
\alias{my_allele_freq}
\title{Major allele freq calculation}
\usage{
my_allele_freq(my_seqs)
}
\arguments{
\item{my_seqs}{A vector of bases}
}
\description{
Calculates the major allele frequence from a set of Nucleotide bases for a given site.
}
\details{
Major allele frequency
}

