% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my_seg_sites.R
\name{my_seg_sites}
\alias{my_seg_sites}
\title{Segregating sites}
\usage{
my_seg_sites(my_seqs)
}
\arguments{
\item{my_seqs}{A dataframe of genomic sequence information, where each row is an individual sequence (haploid).}
}
\description{
A function to locate the segregating sites in a data frame of complete DNA sequecne
}
\details{
Locate segregating sites

A vector of the position of segregating sites
}

