% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my_Fst.R
\name{my_Fst}
\alias{my_Fst}
\title{calculate pairwise fst
Weir and Cockerham 1984
give number of populations, number of individuals in each pop.,
and major allele frequency}
\usage{
my_Fst(r, n, p)
}
\arguments{
\item{r}{number of populations}

\item{n}{number of individuals in each population}

\item{p}{major allele freq in each population}
}
\value{
Pairwise Fst
}
\description{
calculate pairwise fst
Weir and Cockerham 1984
give number of populations, number of individuals in each pop.,
and major allele frequency
}
\references{
Weir and Cockerham 1984
}

