% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/my_pi.R
\name{my_pi}
\alias{my_pi}
\title{Nucleotide diversity}
\usage{
my_pi(my_seqs)
}
\arguments{
\item{my_seqs}{A dataframe of genomic sequence information, where each row is an individual sequence (haploid).}
}
\description{
Calculate pi from all combinations
}
\details{
Pairwise nucleotide diversity for whole sequence DNA
}
\examples{
bases<- c("A","T","C","G")
seq1 <- sample(bases,1000, replace=TRUE)
seqs<- as.data.frame(matrix(NA,nrow=9,ncol=1000))
seqs[1,]<-seq1
for ( i in 2:10){
snps<- sample(bases,3, replace=TRUE)
snp_pos<-sample(seq(1,1000),3)
seq2<- seq1
seq2[snp_pos]<-snps
seqs[i,]<-seq2}
my_pi(seqs)
}

