% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/my_pi.R
\name{my_pi}
\alias{my_pi}
\title{Nucleotide diversity}
\usage{
my_pi(my_seqs)
}
\arguments{
\item{my_seqs}{A dataframe of genomic sequence information, where each row is an individual sequence (haploid).}
}
\description{
Calculate pi from all combinations
}
\details{
Pairwise nucleotide diversity for whole sequence DNA
}

